requirements.txt
python-dotenv
pytelegrambotapi
rich